webservice/requirements.txt
celery~=5.2
redis~=4.3
Django~=3.2
opencv-python==4.6.0.66
scikit-learn
numpy
matplotlib~=3.3
ktrain
tensorflow
djangorestframework~=3.14
gunicorn==20.1.0
psycopg2-binary==2.9.5
imageio
cchardet
drf-chunked-upload
keras_vggface
git+https://github.com/rcmalli/keras-vggface.git